backend/health_record_backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
python-multipart>=0.0.6
cryptography>=41.0.0

# PDF support
pypdf>=4.0.0

# Optional: For scanned PDF OCR support (requires tesseract installed)
# pytesseract>=0.3.10
# pdf2image>=1.16.0
